// src/entities/facility/model/types/response.ts
import { Facility, PageInfo } from '.';

export interface FacilityResponse {
  content: Facility[];
  pageable: PageInfo;
  last: boolean;
  totalPages: number;
  totalElements: number;
  first: boolean;
  size: number;
  number: number;
  sort: {
    empty: boolean;
    sorted: boolean;
    unsorted: boolean;
  };
  numberOfElements: number;
  empty: boolean;
}